// src/domain/enums/ScheduledTaskType.ts
export enum ScheduledTaskType {
  INVOICE_PROCESSING = 'INVOICE_PROCESSING',
  APPROVAL_ESCALATION = 'APPROVAL_ESCALATION',
  APPROVAL_REMINDER = 'APPROVAL_REMINDER',
  PAYMENT_PROCESSING = 'PAYMENT_PROCESSING',
  PAYMENT_RECONCILIATION = 'PAYMENT_RECONCILIATION',
  RECONCILIATION = 'RECONCILIATION',
  RISK_ASSESSMENT = 'RISK_ASSESSMENT',
  COMPLIANCE_CHECK = 'COMPLIANCE_CHECK',
  REPORT_GENERATION = 'REPORT_GENERATION',
  DATA_CLEANUP = 'DATA_CLEANUP',
  BACKUP = 'BACKUP',
  NOTIFICATION_DIGEST = 'NOTIFICATION_DIGEST',
  AUDIT_LOG_ARCHIVE = 'AUDIT_LOG_ARCHIVE',
  SUPPLIER_RATING_UPDATE = 'SUPPLIER_RATING_UPDATE',
}